refactor(node): absorb tools crate into node::hooks module
Remove the standalone crabtalk-tools crate. All hook implementations
(OsHook, MemoryHook, AskUserHook, SkillHook) move into
node::hooks::*. Prompt files move to node/prompts/. The tools crate
had no consumers outside node and bench — bench now imports bm25 from
node::hooks::memory::bm25.
