+++ crates/node/prompts/crab.md
