+++ crates/node/prompts/memory.md
