+++ crates/node/src/hooks/memory/handlers.rs
