+++ crates/node/src/hooks/mod.rs
@@ -1,0 +1,8 @@
+//! Built-in hook implementations — tool subsystems registered on Env.
+
+pub mod ask_user;
+pub mod memory;
+pub mod os;
+pub mod skill;
+
+pub use memory::Memory;
