+++ crates/node/src/hooks/skill/handler.rs
