+++ crates/node/src/hooks/skill/loader.rs
