+++ crates/node/src/hooks/skill/mod.rs
