+++ crates/node/src/lib.rs
@@ -4,7 +4,8 @@ pub(crate) mod builder;
 pub mod cron;
 pub mod delegate;
 pub mod event;
-pub mod hook;
+pub mod hooks;
+pub mod host;
 pub mod mcp;
 pub mod node;
 mod protocol;
@@ -13,9 +14,9 @@ pub mod provider;
 pub mod storage;
 
 pub use builder::{BuildProvider, DefaultProvider, build_default_provider};
-pub use hook::NodeEnv;
+pub use hooks::Memory;
+pub use host::NodeEnv;
 #[cfg(unix)]
 pub use node::setup_socket;
 pub use node::{Node, NodeHandle, bridge_shutdown, setup_tcp};
-pub use tools::Memory;
 pub use wcore::NodeConfig;
