+++ crates/node/src/protocol/conversation.rs
@@ -106,7 +106,7 @@ pub(super) fn stream<'a, P: Provider + 'static, H: Host + 'static>(
                         .iter()
                         .filter(|c| c.function.name == "ask_user")
                         .filter_map(|c| {
-                            serde_json::from_str::<tools::ask_user::AskUser>(&c.function.arguments)
+                            serde_json::from_str::<crate::hooks::ask_user::AskUser>(&c.function.arguments)
                                 .ok()
                         })
                         .flat_map(|a| a.questions)
