--- crates/tools/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "crabtalk-tools"
-description = "Built-in tool implementations for Crabtalk"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-repository.workspace = true
-keywords.workspace = true
-
-[dependencies]
-runtime.workspace = true
-wcore.workspace = true
-
-# crates.io
-anyhow.workspace = true
-schemars.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-serde_yml.workspace = true
-tokio.workspace = true
-tracing.workspace = true
